feat: move the session rollover demo into its own example
The demo landed on top of hello-world-web, which meant the SDK's minimal
getting-started example was also the session dashboard. Those are two
different jobs: hello-world-web should stay the smallest thing that works,
and the session demo needs short windows, buttons and panels that only make
sense if you came looking for sessions.

Splits it out as examples/session-rollover, restoring hello-world-web to
exactly what is on main, and wires the new directory into example:install:all
and the per-example install/start scripts.

Keeps hello-world-web's package-lock.json refresh. The "../.." entry there
mirrors the SDK's own dependency block and had gone stale at version 1.0.3
and the 0.203.0 generation; the refresh records 1.4.0, 0.221.0 and this
branch's core/resources bump. That belongs to the dependency change rather
than to the demo.

Also sorts the import in the demo, which eslint flags under sort-imports.
