+++ packages/honeycomb-opentelemetry-web/examples/session-rollover/.npmrc
@@ -1,0 +1,2 @@
+ignore-scripts=true
+min-release-age=7
