+++ packages/honeycomb-opentelemetry-web/package.json
@@ -48,7 +48,7 @@
     "test:update-snapshots": "jest --config ./jest.config.js --no-cache -u",
     "test:ci": "jest --config ./jest.config.js --ci --runInBand --reporters=default --reporters=jest-junit --no-cache",
     "example:lint:all": "npm run example:install:all && eslint examples",
-    "example:install:all": "npm run example:install:custom-with-collector-ts && npm run example:install:experimental/user-interaction-instrumentation && npm run example:install:hello-world-cdn && npm run example:install:hello-world-cjs && npm run example:install:hello-world-custom-exporter && npm run example:install:hello-world-react-create-app && npm run example:install:hello-world-web",
+    "example:install:all": "npm run example:install:custom-with-collector-ts && npm run example:install:experimental/user-interaction-instrumentation && npm run example:install:hello-world-cdn && npm run example:install:hello-world-cjs && npm run example:install:hello-world-custom-exporter && npm run example:install:hello-world-react-create-app && npm run example:install:hello-world-web && npm run example:install:session-rollover",
     "example:install:custom-with-collector-ts": "cd ./examples/custom-with-collector-ts && npm ci",
     "example:install:experimental/user-interaction-instrumentation": "cd ./examples/experimental/user-interaction-instrumentation && npm ci",
     "example:install:hello-world-cdn": "cd ./examples/hello-world-cdn && npm ci",
@@ -56,13 +56,15 @@
     "example:install:hello-world-custom-exporter": "cd ./examples/hello-world-custom-exporter && npm ci",
     "example:install:hello-world-react-create-app": "cd ./examples/hello-world-react-create-app && npm ci",
     "example:install:hello-world-web": "cd ./examples/hello-world-web && npm ci",
+    "example:install:session-rollover": "cd ./examples/session-rollover && npm ci",
     "example:start:custom-with-collector-ts": "cd ./examples/custom-with-collector-ts && npm start",
     "example:start:experimental/user-interaction-instrumentation": "cd ./examples/experimental/user-interaction-instrumentation && npm start",
     "example:start:hello-world-cdn": "cd ./examples/hello-world-cdn && npm start",
     "example:start:hello-world-cjs": "cd ./examples/hello-world-cjs && npm start",
     "example:start:hello-world-custom-exporter": "cd ./examples/hello-world-custom-exporter && npm start",
     "example:start:hello-world-react-create-app": "cd ./examples/hello-world-react-create-app && npm start",
-    "example:start:hello-world-web": "cd ./examples/hello-world-web && npm start"
+    "example:start:hello-world-web": "cd ./examples/hello-world-web && npm start",
+    "example:start:session-rollover": "cd ./examples/session-rollover && npm start"
   },
   "author": "Honeycomb <support@honeycomb.io> (https://www.honeycomb.io/)",
   "license": "Apache-2.0",
